+++ NEWS.md
@@ -1,8 +1,8 @@
-# LBBNN 0.1.1 (development version)
+# SICNN 0.1.1 (development version)
 * Initial tests added; cleanup for R CMD check portability.
 * Initial CRAN submission.
 
-# LBBNN 0.1.2 (minor bug fix)
-* Fix bug in LBBNN_Linear so layers work properly in convolutional architecture.
+# SICNN 0.1.2 (minor bug fix)
+* Fix bug in SICNN_Linear so layers work properly in convolutional architecture.
 * Updated experiment on convolutional architecture for R journal submission.
 * Added default values to mpm and draws in predict() function.